Renamed files for `JoltCustomShape*Type`
