+++ src/spaces/jolt_motion_filter_3d.cpp
@@ -4,8 +4,8 @@
 #include "objects/jolt_object_impl_3d.hpp"
 #include "servers/jolt_physics_server_3d.hpp"
 #include "shapes/jolt_custom_motion_shape.hpp"
+#include "shapes/jolt_custom_shape_type.hpp"
 #include "shapes/jolt_shape_impl_3d.hpp"
-#include "shapes/jolt_shape_type.hpp"
 #include "spaces/jolt_broad_phase_layer.hpp"
 #include "spaces/jolt_space_3d.hpp"
